+++ extensions/vscode/src/utils/install-bloc-tools.ts
@@ -1,7 +1,7 @@
 import { exec, isDartInstalled } from ".";
 
+export const blocToolsVersion = "0.1.0-dev.10";
 export const installBlocTools = async (): Promise<boolean> => {
-  const blocToolsVersion = "0.1.0-dev.9";
   const canInstall = await isDartInstalled();
   if (!canInstall) return false;
   try {