+++ extensions/vscode/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 6.8.1
+
+- feat: improve bloc tools install/upgrade ux
+- fix: various language server bugs on windows
+- deps: upgrade to `bloc_tools: 0.1.0-dev.10`
+
 # 6.8.0
 
 - feat: add language server